Clarify the purpose of the INSTALL file
INSTALL is a generic document that is provided with autoconf to
document the features of the configure script that it produces.  It is
considered an essential part of any project that uses autoconf.

Some users have confused it for install instructions for Xyce, which
is a huge mistake.  It is unfortunate that the entire autoconf world expects
this file to be called INSTALL when in fact it's just a generic thing
about configure.  It would be better called "configure.doc".

Several releases ago I added text to the top of INSTALL directing
users not to expect INSTALL to document the Xyce build process, and
telling them where to look instead.  I've rearranged and amplfied
that text.