Initial Commit for FIND-WHEN and WHEN Measures for .AC
This is the initial commit for FIND-WHEN and FIND measures for .AC
analyses.  Support for FIND-WHEN and FIND measures for .DC will be
added later.  This commit also changed the error message that is
emitted by netlist parsing when an unsupported measure mode
(e.g., ON_TIME) is requested on a .MEASURE AC or .MEASURE DC line.

SON Bug 1270
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1270