Cleans up checkNodeDevConflicts method, reduces communication
In profiling the 10M device memory circuit from Yale, the checkNodeDevConflicts
method exhibits an excessive amount of cost from communication when run on larger
number of processors.  This was the result of a distributed device name check
that essentially communicated all devices names to every processor in a
round-robin manner.  So, np*(np-1) communications, where np is the number of
processors, when (np-1) communications would have sufficed.  Of course, this
will only look bad on larger numbers of processors.  This check is now rewritten
and memory and communication is reduced.