Removes unused objects and Epetra_Util usage
The Epetra_Util sort method is used in the parallel topology utility
for sorting the external GIDs based on which processor they come from.
To use this utilty, multiple copies are generated, when a much simpler
approach to collecting this information would avoid more than one copy.

In reviewing some code in the graph class for handling external GIDs
and passing them to device objects, usused objects were found and also
unnecessary copies.