Fix bug 1264 SON; remove excess processParam calls
It was observed recently that if TEMP was used on a .DC line as part of
a nested sweep, that Xyce would call processParams and
updateTemperature on every single
sweep step, rather than just the steps in which the TEMP variable
actually changed.  It was also observed that if the TEMP variable was
removed from the .DC line and then added to a .STEP line, that this
problem went away, and updateTemperature was only called when it was
actually needed.

My theory was that for sweeps, the DeviceMgr::setParam function was
being unconditionally called on every single sweep variable.  After
digging thru the code for a few minutes, I observed that this was the
case.

Fortunately the fix was very easy.  For the old style sweeps, (like LIN,
OCT, DEC, LIST) there are various index variables used to keep track of
everything.  I used these index variables to determine if a sweep
parameter was changing or not.  If not, then a boolean is set that
prevents DeviceMgr::setParam from being subsequently called.

This change was pretty easy, and tests are all passing.

As part of the new logic, if the indices do not indicate that a sweep
parameter needs a new value, a new value isn't computed at all.  So, the
old value (which is in a persistent variable) is simply reused.  This is
a very small optimization (as this operation was very cheap) but it
was easy to include.

For the more modern sweeps covered by this function (TABLE and various
types of SAMPLING) I didn't apply this new logic.  The reason being that
for those types, it is normal for all the variables to change at once.