Allow Xyce's buiild to accept Bison 3.5
Shortly after we committed to releasing 7.0, Bison 3.5 started showing
up in package management systems.  It works Just Fine with Xyce, but
configure is explicitly set up only to accept versions of bison we
have *CHECKED* for compatibility with Xyce.

Now that I've checked that 3.5 works, I'm adding it.

I expect we'll get some complaints about this, so we should be
prepared to tell open source users how to fix it in 7.0.