Changes generateOffset to be the maximum GID, instead of a factor of 10
The generateOffset method is used by HB and MPDE to determine which offset
should be used in the generation of GIDs for the block system.  Originally,
this number was the factor of 10 larger than the maximum GID.  This was
confusing and presumably because there may be a need for any processor to
add unknowns (like for phase conditions, etc.).  In practice, we assign any
individual conditions to the last processor.  So, using an offset larger
than the maximum GID is unnecessary.  Furthermore, it requires that any
direct linear solvers that Xyce uses (speaking more to MPDE here) need to
reorder the indices so that they are contiguous.  Otherwise, they often
return an error that the linear system is rectangular or singular.

Ting and I agree this is the best course of action moving forward and it
facilitates the integration of the multi-threaded Basker solver because it
does not have reordering.