Allow Expressions for OBJVAL for WHEN Measures
This commit allows a WHEN measure to use a variable quantity (e.g.,
a solution variable) in an expression on the right-hand side (also
called the OBJVAL in the source code) of its WHEN clause.  An example
is:

  .MEASURE DC whenExp2 when v(1)={v(2)+1.2}

Previously, that OBJVAL would always evaluate to 0 in the example
shown above.

SON Bug 1273
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1273