Removes KLU_REINDEX from linear solver options
The need for reindexing should be addressed through other means.
In particular, the current need for this option is because of an
offset issue that has been addressed for block linear algebra.  I'm
removing the option so we don't get comfortable using it, as it will
limit our abilities to integrate newer direct solvers.