Fix for bug 807 SON, which was a BSIM4 bug
The title for bug 807 SON is: "Investigate BSIM4 convergence problems with rgatemod=1"

It had been observed that when using the rgateMod=1 option in the BSIM4,
that the model had convergence problems.  Several circuits that ran fine
with rgateMod=0 had problems with rgateMod=1, and died with
time-step-too-small errors.

The PTM PDK model cards (from ASU) often use the BSIM4 with rgateMod=1
and many users have been trying to use those model cards with Xyce.  So,
this became a much higher priority issue recently.

This commit fixes the problem.  Here is a brief explanation:

The problem was with the vgb voltage drop variable. (i.e., gate-to-bulk voltage).
And, it was only with the limiter details of that voltage drop, specifically the
“vgb_orig” variable.  It wasn’t being set up properly, and was (I think) always
zero.  As a result, it often looked like large limits needed to be applied for
that voltage drop, and that was inconsistent with what was actually happening
with the solution.

This variable was only used by the capacitors, so it only manifested in
transient.  This mistake would not have mattered during the DCOP.

Fixing this required a single line of code be added to the BSIM4, to
properly estabilsh vgb and vgb_orig prior to limiting being computed.
Once I added this line, all the problematic circuits attached to bug 807
ran easily to completion.