Fix SON Bug 1275 (Sorting of PAUSE and SIMPLE Breakpts)
This commit explicitly enforces a sort order/priority for SIMPLE
and PAUSE breakpoints in the BreakPointLess struct.  This should
fix the bug that occurred when a PAUSE and SIMPLE breakpoint occurred
at the same time during Mixed Signal Interface usage.  PAUSE
breakpoints are now considered < SIMPLE breakpoints for that case.

SON Bug 1275
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1275