Improve Interpolation for FIND-WHEN and WHEN Measures
This commit improves the interpolation algorithm for the FIND-WHEN
and WHEN measures for all three measure modes (AC, DC and TRAN).
Previously, that interpolation assumed a fixed "target value" for
the WHEN clause.  So, it worked for WHEN V(1)=A where A was a fixed
value.  However. the interpolation could be inaccurate for the case
of WHEN V(1)=V(2) where the "target value" (of V(2) in this case)
was a varying quantity.

SON bug 1273
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1273