Another Improvement to FIND-WHEN and WHEN Measures
This commit makes another improvement to the FIND-WHEN and WHEN
measures for all measure modes (AC, DC and TRAN).  The previous
algorithm for determining whether the dependent variable had crossed
the target value in the WHEN clause did not always work well when
both sides of the equality in the WHEN clause (e.g., the case of
WHEN V(1)=V(2)) could be varying.  This new approach covers that
case.

SON Bug 1273
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1273