Improve Error Handling for Mixed Signal Interface
This commit makes a user-attempt to have requestedTime <= (the current
simulation time), in a simulateUntil() call by the Mixed Signal
Interface, be a Xyce error that causes Xyce to abort.  Previously,
the case of requestedTime < (the current simulation time) would
cause the Xyce simulation to run to completion, with a warning
message.  The case of requestedTime == (the current simulation time)
would cause Xyce to get stuck in a series of "time-step too small"
errors.  Both cases are now a Xyce error.

SON Bug 1075
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1075