Updated to handle subcircuit names (i.e., w/ ':')
I hadn't bothered with this use case until now.

I had to modify the regular expression for "NODE", which is used to lex
out voltage node names,  and also for "ID" which lexes device instance
names for current variables.  Both were modified to include the colon
':'.

Of course, other simulators use different deliminters for this, so
possibly other delimiters should be added to the regex for these.  But
for now, I am just worrying about native Xyce.