Added unit test related to SENS/res.cir test
This unit test incldues the exact same objective function as res.cir.

My initial attempts at applying the new expression library are inside of
the parameter sensitivity capability.   So, I am attempting to run all
the SENS tests.  This was the first one, which initially was failing, so
I decided to debug it (partially) with a related unit test.