Fix use case for Xyce, same obj func variable 2x
At least one of the regression tests has objective functions like this:

{V(B)*V(B)+3.0}

Before this fix, the newExpression library integration with Xyce would
screw this up.  I had to ensure that everything in the names vector was
unique.  ie, if V(B) appears 2x, it should only be set in the names/vars
vector 1x.