Add specializations for "greaterThan" template
The interpolator classes are templated, and can thus use a variety of
data types.  I also use these interpolators in the "tableOp" class of
the new expression library.

Prior to using N_UTL_Inteprolator in the new expression library, it had
never been used on complex numbers.  The one thing I had to change to
make it support std::complex was to modify a handful of greater than
operations to only act on the std::real part of the number.

This worked fine when I was only attempting to use either double or
std::complex<double>, because both of these data types do just fine with
std::real.   So for a long time with my stand-alone version of
newExpression, I thought this issue was resolved.

However, when I tried to integrate this into Xyce a while ago, I
discovered that this wouldn't work, because std::real didn't play well
with sacado types, at least not directly.  There are some workarounds
for this; notably Eric Phipps created some sophsitcated code to make
sacado and std::complex play well together.  So I considered applying it
hear.  However, it has been a few years and I didn't remember enough
about Phipps code to be comfortable with it.

A much quicker work around was to create a templated "greaterThan"
function, which by default simply used "left>right", but that had
specializations for complex numbers (std::real(left)>std::real(right)).

It turns out that my first specialization was not adequate.  I needed to
also create one for "const" std::complex<double>, in order for my
complexParserUnitTest to compile.  This commit implements this additional
specialization, so it now compiles again.  I guess I had not attempted
to compile/run the complex parser unit tests prior to now (since
switching to working inside of Xyce, primarily).

Anyway, seems to work again.