Two tests for parameters names I and/or V.
I discovered yesterday that the parser for the new expression library
had a flaw, and that flaw was that it could only process "I" or "V" as
solution variable currents or voltages.

But, the old parser allows these characters to be parameter names as
well, as long as they aren't followed by parenteses with arguments.  So,
for example:

In the expression {RES*V}, both RES and V are parameters.

In the expression {RES*V(1)}, both RES is a parameter, but V is a
voltage node variable.

I had written the parser to only handled the second case.

This commit contains two new unit tests of the parser that will only
pass when this is fixed.