Fix for parameters named I and/or V
I discovered yesterday that the parser for the new expression library
had a flaw, and that flaw was that it could only process "I" or "V" as
solution variable currents or voltages.

But, the old parser allows these characters to be parameter names as
well, as long as they aren't followed by parenteses with arguments.  So,
for example:

In the expression {RES*V}, both RES and V are parameters.

In the expression {RES*V(1)}, both RES is a parameter, but V is a
voltage node variable.

I wrote the lexer/parser to only handle the second case.

This commit updates the expression lexer and parser to handle this use
case.  The change involved the following.

(1) getting rid of two tokens, TOK_VOLTAGE and TOK_CURRENT, and
replacing them with the more generic (and already in use) TOK_WORD.

(2) modifying the voltagenode and currentdevice "states" so that if they
encounter an unexpected character (the regex for this is simply "."
after all the rules for specific patterns), they will go back to the
INITIAL "state", and prior to doing so, will push the unrecognized
pattern back onto the stack so it can be processed using the rules in
the INITIAL state.

(3) updating the parser *yxx file to accomodate those changes.

Making those changes now means that if the lexer encounters "V", it will
first attempt to resolve it as a voltage node such as V(1), but if
doesn't find the right pattern after the V, it will give up and assume V
is a generic parmaeter.  Same with I.   The unit tests I created to
test this capability now pass.