Add cmd line toggle for new/old expression lib
This is convenient for testing.  It isn't used much at this point, just
in the sensitivity class.  That usage is included in this commit.

Note: the sensitivity class changes include a few changes to debug
output, and some of that will need to be rolled back later as they
output unconditionally.