Make possible for paramOps to be set as vars
This also includes improvements to some of the debug output.

I originally wrote the paramOp class (and the rest of the expression
library) so that it was not subject to differentation, unlike the
voltageOp and currentOp classes.  By default, much of the actual usage
of the expression library in Xyce will only need derivatives with
respect to voltages and currents.

However, the old expression library API has the ability to call a
"set_var" function, (with a string name of an expression field as
argument) which renders (potentially) params as variables which must be
set and/or differentiated.

This commit includes a small new feature to enable the same capability
in the new expression library.  The paramOp class now has an "isVar"
boolean variable, which can be set, unset, and queried.