update to allow differentiation of params
The new expression library was written to differentiate expressions with
respect to voltages and currents, by default.  In principal, it can
differentiate anything, but I didn't initially write code to allow for
this.  For params and global_params, it seemed wrong to unconditionally
do the work of computing their derivatives, as they were often not
needed by Xyce.

This commit updates the code to conditianally allow parameters to be the
subject to differentiation.  They will still not be differentiated by
default, but it is possible now if the user requests it thru the old API
function calls.

In general, I'm trying to not modify the newExpression API very much,
and to have all the stitching between Xyce and newExpression be thru the
N_UTL_Expression class, which serves as an interface.

This commit also includes an initial cut at supporting the use of .FUNC
in different expressions.  This doesn't quite work correclty yet, but it
is close.  With the old expression library, .FUNC is supported via
string substitution, and once that substitution is complete, then the
new, resulting function has no awareness of the .FUNC that it once
called.

For the new epxression library, I don't want to do this at all.
However, in order to plug in the new library with a minimum of mods to
Xyce, I've been hacking some provisional interface code into the
N_UTL_Expression library.  For .FUNC support, what I'm attempting to do
is when the .FUNC string is passed into the expression, instead of
string substitution, I simply allocate a new expression and then attach
that expressions node to the tree of the original.

This still will mean a lot of redundant behavior, and I intend to do
away with this approach once I've sorted out other details.  But I am
doing it for now to enable other types of debugging, etc.