Yet another fix for derivatives thru funcs
In my previous commit, I thought I had derivatives thru .func calls
sorted out, but I was wrong.

This worked:
.func f(x,y) {x+y+10*V(C)}
{f(V(A),V(B))}
where I want the derivatives of the second expression w.r.t. V(A), V(B) and V(C).

But this did not:
.func f(x,y) {x+y+10*V(A)}
{f(V(A),V(B))}
where I want the derivatives of the second expression w.r.t. V(A) and V(B).

The reason the second case didn't get it right was because I made an
indexing mistake, and V(A) was being double-counted.  Once as a free
standing variable, and again as an argument.

So, the derivative of the second example w.r.t. V(A) should have been
11.0.  Instead it was 21.0.

This is fixed now and there are tests demonstrating it, in the unit
tests.