Make complex and double tests more similar
I have endeavored to have the parserUnitTest.C (in which the underlying
type used by all the templates is a double) and the complexParserUnitTest.C
(in which the underlying type is std::complex<double>) programs
basically test the same things, except for when it simply isn't
possible.

However, sometimes the complex number tests fall behind the "double"
tests.  I recently spent some time bringing the complex tests up to date
and the result of that work is in this commit.

This includes complex versions of all my new tests of .func usage, and
the propagation of derivatives thru .func calls.

I also noticed a few small inconsistencies in the parserUnitTests
program, so I fixed those as well.  But they represent maybe 1% of the
changes in this commit.  The meaningful changes are all in
complexParserUnitTest.C.