Improve use .param and .global_params in objfunc
With this commmit, the bug645_Param.cir and bug645_Global.cir tests now
pass.  They were the last two sensitivity tests that were not working.
Now, ALL the sensitivity regression tests pass using the new expression
library.

The paramOp class needed some refinements to better support all the
different ways in which parameters are used.  The paramOp class is used
to support both .param and .global_param, so it needed a little more
flexibility.  Also, sometimes derivatives are requested for parameters
and sometimes derivatives are requested for subordinate variables.
(i.e., if a .global_param includes voltages, then you'll want the
derivatives with respect to the param itself and also with respect to
the voltages).  So, this took some thinking and refactoring.

Anyway, the existing regression tests now all pass for sentivities.
This has also driven the creation of a lot of new unit tests, for cases
I hadn't considered before.

other than the paramOp class, the rest of this work was mostly in the
interface between Xyce and newExpression, which is mostly in the
N_UTL_Expression file and also in the xyceExpressionGroup class.

Next I will be moving on to applying the new expression library in
another part of the code.  Not sure which part yet.

As I am implementing this by mimicing the old expression library API,
there is a certain amount of kludgy-ness I'm having to live with.
Longer term I don't want the old API to be the way things are done, but
for now, it is forcing me to consider all the ways the old library is
used.