Fix for two memory leaks
Last week I updated the Expression lexer/parser so that it could handle
parameters named "I" and "V", and not always assume that those tokens
automatically meant currrent or voltage variables.

In doing that update, I added two more uses of the TOK_WORD token, which is
allocated as a string.  It then has to be deallocated later, which I neglected
to do.  Fixing it wasn't difficult; I just had to add 2 "delete" commands to
the *yxx file.