Update to compile on valinor (RHEL7 gcc5.3)
Prior to yesterday, I had not attempted to compile any of
this on valinor (RHEL7, gcc5.3) since migrating the new expression
library into the a fork of Xyce, under the UtilityPKG.

The new expression library required modification to integrate
into Xyce.  Partly, when compiling it in stand-alone mode, I had copied small
bits of source code from Xyce, and kept them as local copies in the
newExpression src directory.   This included things like a copy of
N_UTL_BreakPoint, and some features of the ExtendedString.h file, etc.

That all had to go when integratging into Xyce so there wouldn't be duplicated
code, and the code had to be modified to the original Xyce.

There were other various changes.  None of them too significant, but it
took a little time.

Anyway, I had made all of that work on OSX, but hadn't bothered yet with
linux, RHEL7, which is my desktop machine.  yesterday, I decided I needed
to test it against valgrind, so that required that I get this going again
on linux.    With this commit, all the unit tests now compile on linux,
inside the new Xyce fork.

On linux, I had to change the order of the libraries
links at the final link.   That was kind of annoying.

Currently I'm not attempting to use CTest or anything like that.  I'm still
using simple hand-written local makefiles, that are very specific to my
machines.  This is not designed (for now) to be very portable.  That will come
later as CMake build system for Xyce comes on line.

Note, also, that in my recent changes I was only running the parserUnitTest and
complexParserUnitTest unit test programs.  I had not run the astUnitTest
program in a few days, and it looks like I broke many tests in it.  So, if you
try to build and run astUnitTests, you will see a lot of failures.  I am
working on fixing it now.