fix to remove use of yyless
A couple of weeks ago I fixed the issue of using "I" and "V" as
parameter names.  Before that, they had to be the prefix for voltage and
or current variables.  After my fix they could be either.

Today I was fixing something else, and it occured to me that my previous
fix may have been the cause of my new problem.  This turned out to be a
red herring.  However, in the process, I discovered that I don't think I
was confident in my use of the yyless macro command.   So, I chose to
find another solution to the original problem, that didn't involve
yyless.  This commit contains that new method.