Improvements to get newExpression working w/Bsrc
This commit contains a bunch of changes to flesh out the
N_UTL_Expression interface to newExpression.  Previously, I had only
implemented enough functions to make the sensitivity tests pass, and
this was a small subset of the API.

This commit doesn't score a touchdown, but it moves the ball down the
field another 10 yards or so.  With the changes in this commit, I'm able
to unconditionally turn on the new expression library, which means it is
used for ALL uses of expressions, and then pass about 1/2 of the Bsrc
tests.

Of the ones that are failing, some are failing for known reasons (like I
still haven't implemented an API function) and some are failing for
unknown reasons (like a suspected memory bug).  The latter possibility
is why I am doing this commit now, so I can run this thru valgrind on
another computer.

Changes in this commit include:
 - realizing that the copy constructor should *not* be re-parsing
expressions.  That was causing weird problems.
 - more work on get_names
 - implement get_type
 - implement all forms of evaluate and evaluateFunction
 - implement set_vars

... and various other things.