Add unit tests related to the ascth.cir test
One test that currently fails when I used newExpression in Xyce is the
ascth.cir test.  This test doesn't seem to be one that *should* fail, as
it uses very simple arithmetic functions (sinh, asinh, cosh, acosh), and
those functions were implemented in newExpression.

I wanted to double check that the core code for these functions would
work with the exact arguments used in ascth.cir on the first DC sweep
step.  It looks like that code works.  The new tests pass.  So, the
error is somewhere in the interface from Xyce to N_UTL_Expression to
newExpression.