Fix Array Indexing Errors in Touchstone Input Parsing
This commit fixes several array indexing errors in the handling of
Touchstone input files for the YLIN device.  The handling of S(i,j)
vs. S(j,i) is now correct, which is important for non-symmetric
S-, Y- or Z-parameter matrices.  This commit also fixes the handling
of the optional [Two-Port Data Order] line.

SON Bug 1215
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1215