Add initializer for matrixMarketFormat_ var
This only matters when running in debug mode and are outputting
matrices.  There is a flag that sets if the outputted matrix from the
nonlinear solver uses matrix market format or not.    It was not being
initialized in the constructor of N_NLS_NLParams.   It had get/set/reset
functions and the reset function set it to "false", so I chose to
initialize it to false as well.  The code was behaving (for me) as
though its initialization was false, so this preserved that behavior as
well.