Add support for atan2
This is supported in the old expression library, but I had neglected to
support it in the new one.  It is supported now.

I think my support may be better than that of the old library, in that
I think I'm properly handling all four quadrants, and I think the
old library did not.

In the parser, I've also added a little bit of book-keeping for specials
such as "TIME", "TEMP, etc.  This is needed to better support dependent
parameters in Xyce.