Add support for replace_name and other fixes
This commit includes changes from several days of work, so it is
hard to summarize everything.  This is work driven by trying to get the
Xyce regression tests specified by this set of tags to run:

+nightly+serial+bsrc?klu

That adds up to 127 tests.  As of this writing, 40 exit with error, 11
run but fail comparison, and the rest (76) pass.

This commit includes the following:
(1) support for the replace_names function

(2) support for providing XEXP_SPECIAL type in the get_names function.
I had not set this up before, and it was preventing Bsrc's that had
time-dependent (but not solution dependent) expressions from running
correctly.

(3) support for get_break_time().  I really don't like this function, and
I have not (yet) supported it in the same spirit as the way it is
implemented in the old expression library.  I have implemented the new
expression library so that traditional discontinuous sources such as
spice_pulse have well defined breakpoints.  The old library would
attempt to find them on the fly, by "solving" for them.   This "solving"
technique is interesting and useful (perhaps) for functions like STP,
when STP has "time" as the input.  But for well-defined static sources,
it makes no sense.  Anyway, to support it here I simply pull out the
breakpoints vector from the new expression AST, and then return
the one (if any) that is closest to the current time, and also
in the future.  I have NOT implemented a quasi-Newton-ish "solve".

(4) failsafes for Nan's and infs.  These probably aren't perfect, but
they catch a lot of cases.  Previously I didn't have enough protection
for this.

(5) support for atan2 (this was actually in a previous commit, I think).

(6) support for synonym for atan, arctan.  There are probably going to
be more synonyms I need to add to the parser.  Fortunately, this takes
roughly two seconds of work.

Known things that still don't work:
(1) Use of expressions on the print line.  This doesn't yet work because
I haven't implemented the "getSymbolTable" function from the old
expression library.  The name makes this sound complicated, but looking
at it more closely I am not too afraid of it.  So I will do this next.

(2) proper resolution of external parameters and .funcs.  I had taken
care of this (I thought) for the use cases in .sens.  But I apparently
didn't populate enough of the "get_names" and other API functions to
support this type of resolution in the IO package of Xyce.  So that is
coming soon as well.