Add suppoort for POLY
I had neglected to support the POLY capability.  Partly because I never
use it, partly because I don't like it, and partly because I didn't
really understand the implicit ordering used by this feature in Xyce
(which is part of why I don't like it)

But, I have now implemented it, and it works.  Most of the Bsrc
regression tests that use POLY will now pass.  The ones that don't pass
will fail for reasons unrelated to POLY support.

As of this commit, 78 out of 127 tests pass using these tags:

+nightly+serial+bsrc?klu

One other significant feature I haven't properly supported yet:
derivatives for the case of V(a,b).    This issue is part of why the
polyg.cir test still fails.  That will come next.