fix for derivatives of V(A,B)
My original idea for V(A,B) was to have the voltageOp class handle the
differencing.  Unfortunately, this made my design for handling
derivative indexing hard to set up.  My derivative index design assumes
that any operator will have at most a single derivative index to worry
about.

For the case of V(A,B), there would need to be two separate indices.

It then struck me that it would be much simpler to handle the V(A,B)
case by just creating a sub-expression during parsing, which would
consist of a voltageOp for V(A) and another voltageOp for V(B), and
these would be arguments to a binaryMinuOp operator.  And *that*
operator would be the one that represented V(A,B).  That way, the
indexing could still be (at most) a single index for each operator.

This unfortunately breaks one of my unit test, ddx8.  But looking at
that test, I don't know if the test is an important use case or not.
The test is of this:

ddx(sin(V(A,B),V(A,B))

in other words, the equivalent of ddx(sin(V(A)),V(A)), but the variable
is a voltage difference instead of a voltage.

An equivalent way of doing it, which would work is:
.global_param p1 {V(A,B)}
ddx(sin(p1),p1)

Anyway, leaving ddx8 broken for the moment until I decide what to do
with it.

With this fix, one more regression test passes, and a bunch that were
exiting with error now run, but fail.   So, out of 127 tests, that use
the "+serial+nightly+bsrc?klu" tags:

There are 127 tests
12 Tests FAILED
36 Tests EXITED WITH ERROR