Fix for log = log10 instead of log = ln
I had assumed, incorrectly that "log" was a synonym for std::ln.  It is
not, at least in Xyce.  In Xyce it means log10.  It probably means
natural log in Hspice, or something like that.  Bleah.

So, that is fixed, and two unit tests (for both complex and double) are
updated to reflect this change.

In this commit, I also correct a couple of things (the
newExpression::getTime function and also the new atan2Op class) to work
properly with complex.

Also, a few of my new parser tests for doubles are now also implemented
in the complex parser tests.

With the fix to log=log10, there is now one more regression test that
passes with the tags "+nightly+serial+bsrc?klu", using the new
expression library.  So we have:

There are 127 tests.
11 Tests FAILED
36 Tests EXITED WITH ERROR

(ie, 80 out of 127 pass)