Support for pwrs operator
I had created a pwrs operator, but hadn't implemented it correctly.

The "val" function had a sign error, and I hadn't yet implmeneted the
"dx" function.

Also it was not unit-tested.  So, not suprisingly, it didn't work.  It
caused one of BINARYOPS regression tests to fail.  It is fixed now, and
I have created a bunch of unit tests on the AST level.

I should also make some parser level unit tests for it but have not done
that yet.

Also, created a test for the unaryMinusOp class.  I hadn't done that
yet.  This operator is implicitly tested in the parser unit tests, but
had zero tests in the AST unit tests.  Now there is one stand-alone, and
it is also used by some of the PWRS tests.

Now, of the Xyce regression tests for
taglist="+nightly+serial+bsrc?klu", we have:

There are 127 tests
10 Tests FAILED
36 Tests EXITED WITH ERROR

Knocking them out, one at a time.