Initial stab at symbol table API function
Some of the remaining failing tests were failing because expressions on
the .PRINT line were not supported yet.   The main thing was missing was
the getSymbolTable function.

This commit adds a first cut of this funciton.  It is very similar to
the get_names function, which is used by much of the rest of the code to
set up expressions.  The main difference is that it allows for more
detailed information about different variables, etc to be provide back
to the calling code.  In particular, it looks like it was set up to
better support things like lead currents in expressions.

I haven't yet bothered with things like lead currents in the new
library, but it will probably be something I tackle soon.  I don't think
it will be that hard to add it, but I haven't tried yet.

Also in this commit, I did some debugging of the new POLY keyword.  The
sandler28 test has been failing, despite my (in theory) supporting all
the expression usage required by that test at this point.  The reason it
was failing is this POLY statement from one of the Bsrc's:

{POLY(5) I(VB) I(VC)  I(VE)  I(VLP) I(VLN)  0 10.61E6 -10E6 10E6 10E6 -10E6}

At first I found this a bit mysterious, but eventually I figured out
what was going on.  I don't do any special tokenizing inside the POLY
statement; I just let all the default lexing happen for fields such as
I(VB) and -10E6.  I thought this would be fine, but it didn't occur to
me that the epxression library would see these two fields:

10.61E6   -10E6

as a binaryMinus operation, resulting in a single coefficient
of (10.61E6-10E6), rather than rather than two fields/coefs,
one of which has a unaryMinus in front of it.    So, if I change
the POLY statement to this:

{POLY(5) I(VB) I(VC)  I(VE)  I(VLP) I(VLN)  0 (10.61E6) (-10E6) (10E6) (10E6) (-10E6)}

(ie, with parens around each number)  then the new expression library
parses it correctly and gets the right answer.

Anyway, I know how to fix this, I think.  I'll have to do some
POLY-specific context lexing, and then modify the parser to use the
poly-specific tokens.

At this point, the taglist="+nightly+serial+bsrc?klu" tests yield the
following:

There are 127 tests
4 Tests FAILED
36 Tests EXITED WITH ERROR