Further refinement of POLY.
Some of the regression tests using POLY were failing yesterday,
including SANDLER28.  The  reason was I wasn't being smart about
lexing/parsing things like this:

POLY(1) V(1) 0.0 -1.0

in which there are two coefficients:  0.0 and -1.0.

The lexer/parser wasn't smart enough to realize that the minus sign in
front of the 1.0 should be treated as a unary minus, rather than a
binary minus.  So, it parsed as though the final argument was (0.0-1.0),
or in other words, for the above POLY statement it believed that there
was a single coefficient and that single coefficient should be -1.

I've added a bunch of context specific logic to the lexer and a bunch of
new token types so that tokens after a POLY statement are handled
differently.  This fixes the use case above.

Now, when I run Xyce_Regression using the newExpression library using
--taglist="+nightly+serial+bsrc?klu" the results are as follows:

There are 127 tests.
3 Tests FAILED
35 Tests EXITED WITH ERROR

this means that 89 out of 127 tests pass.