More refinements of POLY
POLY needed a little more refinement, still.

I have added more lexing/tokenizing to handle the use case of using
parens around the coefficients of poly.  Now this works:

POLY(2) V(1) V(2) 0.0 ((3.0)) {4.0}

whereas it did not before.

I'm also working to make it possible to have the coefficients be
parameters.  I have created one unit test which tests out this scenario.
But the code can't pass it yet.  At the moment the code is hardwired for
the coefs to all be simple numbers.