Support Per-Port Short-Circuit Currents in YLIN Device
This commit adds support for parsing a non-standard extension
to the Touchstone 2 file format to the YLIN device.  It adds
additional columns in the [Network Data] section of that file,
for each frequency, for the per-port short-circuit currents.
This is being done to support coupling to EM analysis codes.

SON Bug 1215
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1215