fix a few bugs, & refactor .func
I've mostly been working on refactoring the handling of .func in the new
expression library so that it is easier to use it behind the old
expression library API.  However, along the way, I uncovered a number of
little bugs that needed fixing as well.

Bugs first:
 - The newExpression default constructor (which takes no arguments)
needed to initialize the few true pointers that are used by this class,
and are deleted in the destructor.  In unit testing, I mostly don't use
this constructor, but in Xyce usage I have to.  As this constructor
didn't initialize these things, it got dangerous when it tried to delete
things that had not been allocated.

- Several single letter operators (P,R, and M) were being tokenized in
such a way that they could not be used as parameters.  Many regression
tests use "R" as a parameter.  So, I had to fix the *l and *yxx files to
be more flexible with these operators.  Now, they can be used as params
and there are simple unit tests that address this directly.

- in a few other places in the code (mostly newExpression.C) I added
extra if-statements so that the code would not attempt to use an RCP
that didn't actually point to anything.

Func refactor:
this is still a work in progress.  There are a few things I'm trying to
address better.  One is that when a .func is declared in Xyce.  Ie,
something like:
.func abc(x,y) {3.0*y+2}

Xyce uses the expression library to parse everything.  So, it feeds
"abc(x,y)" to the expression library, and then calls "get_names" to pull
out the prototype arguments, which will be a std::vector<std::string>
object containing {"x","y"}.    It is important that this vector be in
the order that the user declared it.

In a completely separate expression, it then feeds the expression library
"{3.0*y+2}", and parses it.  At some point it calls "order_names" on the
second one, to inform it that both "x" and "y" are inputs to it, and
that when it is called, the values will come in a std::vector<double>
that are in that order.

I had not considered the use case where newExpression would be parsing
the function prototype (ie, abc(x,y), or the LHS of the above .func),
and bein used to get x,y.  As it turns out, newExpression could parse
it, but returned the arguments in the wrong order because I was storing
them in a std::unordered_map.

To address this I've done two things. (1) I've changed newExpression so
that parameters are no longer stored in an unordered_map, but are
instead stored in a vector.  (2) I've also created new functions to
specifically support this use case, where the calling code knows that
this is a function prototype.    The new functions may not stick around;
I don't think I need both of these fixes.  But for now, they are both
there.

One other thing in this commit; I had mistakenly put a
"resolveExpression" function call into the "N_UTL_Expression::replace_func"
function.  This was too early.  I've now made it so that when
"evaluate" or "evaluateFunction" are called, newExpression will check
and see if "resolveExpression" has been called yet, and if not, do it at
that point.

With these recent changes, Xyce_Regression does this for
taglist="+nightly+serial+bsrc?klu" :

There are 127 tests.
19 Tests FAILED
16 Tests EXITED WITH ERROR

meaning that now 92 out of 127 pass.